refactored app root state and reducers
