+++ src/app/store/store.reducer.ts
@@ -1,0 +1,10 @@
+import { ActionReducerMap } from "@ngrx/store";
+
+import { IStore } from "./store.model";
+import { AuthReducer } from "../auth/store/auth.reducer";
+import { shoppingListReducer } from "../shopping-list/store/shopping-list.reducer";
+
+export const appReducer: ActionReducerMap<IStore> = {
+  authentication: AuthReducer,
+  shoppingList: shoppingListReducer,
+};
